+++ app/src/main/java/com/example/tp5/model/ForecastListResponse/ForeCastListResponse.kt
@@ -2,7 +2,7 @@ package com.example.tp5.model.ForecastListResponse
 
 data class ForeCastListResponse(
     val city: City,
-    val cnt: Int,
+    val cnt: Double,
     val cod: String,
     val list: List<ForecastListItem>,
     val message: Double
